refactor: update kibisis icon component and move icons to domain-level directory
