--- src/extension/components/AnimatedKibisisIcon/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AnimatedKibisisIcon';
