+++ src/extension/components/ModalAssetItem/ModalAssetItem.tsx
@@ -3,7 +3,7 @@ import React, { FC } from 'react';
 
 // components
 import AssetDisplay from '@extension/components/AssetDisplay';
-import WarningIcon from '@extension/components/WarningIcon';
+import WarningIcon from '@extension/components/icons/WarningIcon';
 
 // constants
 import { DEFAULT_GAP, MODAL_ITEM_HEIGHT } from '@common/constants';
