+++ src/extension/components/icons/AlgorandAssetIcon/AlgorandAssetIcon.tsx
