+++ src/extension/components/icons/AlgorandAssetIcon/index.ts
