+++ src/extension/components/icons/AssetIcon/index.ts
