+++ src/extension/components/icons/ImportAccountIcon/index.ts
