+++ src/extension/components/icons/InformationIcon/InformationIcon.tsx
