+++ src/extension/components/icons/InformationIcon/index.ts
