+++ src/extension/components/icons/KibisisIcon/KibisisIcon.tsx
@@ -1,0 +1,13 @@
+import { Icon, type IconProps } from '@chakra-ui/react';
+import React, { type FC } from 'react';
+
+const KibisisIcon: FC<IconProps> = (props: IconProps) => (
+  <Icon viewBox="0 0 1024 1024" {...props}>
+    <path
+      d="m 266,481 c 5.03894,0.94976 8.73819,2.75683 13.1875,5.25 0.78883,0.4411 1.57765,0.8822 2.39038,1.33667 8.96676,5.13492 17.19057,11.19025 25.42212,17.41333 0.80051,0.59909 1.60102,1.19818 2.42578,1.81543 2.38393,1.78414 4.76191,3.57581 7.13672,5.37207 0.73291,0.55277 1.46582,1.10553 2.22095,1.67505 0.67087,0.51329 1.34175,1.02658 2.03296,1.55542 0.5924,0.45141 1.18481,0.90283 1.79516,1.36792 C 324,518 324,518 325,520 c 0.0988,1.7937 0.13283,3.5911 0.13701,5.38751 0.005,1.15082 0.0102,2.30164 0.0154,3.48733 5e-4,1.27526 0.001,2.55052 0.001,3.86442 0.004,1.35991 0.009,2.71981 0.0133,4.07972 0.012,3.7501 0.0176,7.50018 0.0222,11.25029 0.006,4.03974 0.0176,8.07946 0.0283,12.11919 0.0245,9.76422 0.0392,19.52844 0.0521,29.29268 0.004,2.75764 0.008,5.51528 0.0118,8.27292 0.0252,17.12106 0.0473,34.24213 0.0597,51.3632 0.003,3.96759 0.006,7.93517 0.009,11.90276 7.5e-4,0.98614 0.002,1.97228 0.002,2.9883 0.0127,15.96873 0.0459,31.93734 0.0875,47.90601 0.0424,16.39208 0.066,32.78409 0.0712,49.17622 0.003,9.20455 0.0152,18.40893 0.0481,27.61343 0.028,7.8408 0.039,15.68139 0.0278,23.52223 -0.005,3.9993 -0.002,7.99816 0.0225,11.99739 0.0225,3.66576 0.0217,7.33077 0.003,10.99654 -0.004,1.94506 0.0163,3.89013 0.0378,5.83507 -0.087,9.64684 -2.45593,16.58112 -8.90007,23.81979 -7.40888,6.68221 -15.09657,8.02144 -24.86718,7.57031 -8.00038,-0.91755 -13.98885,-4.62411 -19.41407,-10.48437 -5.02685,-6.71139 -6.59462,-14.03021 -6.59179,-22.27652 -0.003,-1.92209 -0.003,-1.92209 -0.006,-3.88301 0.003,-1.41983 0.006,-2.83966 0.009,-4.25949 -4.3e-4,-1.50987 -10e-4,-3.01974 -0.003,-4.52961 -0.002,-4.15028 0.002,-8.30055 0.007,-12.45083 0.004,-4.47507 0.003,-8.95014 0.002,-13.42522 -5.3e-4,-7.75381 0.003,-15.50761 0.008,-23.26142 0.008,-11.21074 0.0107,-22.42147 0.012,-33.63222 0.002,-18.18821 0.009,-36.37643 0.0183,-54.56464 0.009,-17.66927 0.0162,-35.33854 0.0205,-53.00781 4e-4,-1.63341 4e-4,-1.63341 8e-4,-3.29982 10e-4,-5.46283 0.003,-10.92567 0.004,-16.3885 C 265.96004,571.65456 265.97766,526.32728 266,481 Z M 506.67725,-0.39111 c 3.82279,0.0161 7.64452,0 11.46728,-0.019 23.6646,-0.021 47.15369,2.01822 70.48047,6.03516 0.73038,0.12372 1.46077,0.24744 2.21329,0.37491 C 678.86191,21.0532 760.23062,61.76909 826,122 c 0.51063,0.46551 1.02126,0.93103 1.54736,1.41064 8.90964,8.12638 17.73455,16.27576 25.58936,25.44873 2.15215,2.47249 4.38309,4.86334 6.61328,7.26563 5.24155,5.75543 10.03919,11.7925 14.77734,17.96533 1.50974,1.95775 3.0367,3.90109 4.56641,5.84326 64.85172,82.84956 93.28373,182.36299 96.14453,344.61719 0.004,2.38151 0.008,4.76302 0.0123,7.14452 0.0189,11.16291 0.0332,22.32581 0.0416,33.48873 0.01,12.80277 0.0364,25.60537 0.0765,38.40808 0.0302,9.97279 0.045,19.94552 0.0483,29.91835 0.002,5.92726 0.0112,11.85434 0.0365,17.78156 0.0233,5.55477 0.0275,11.10923 0.0172,16.66404 -4.9e-4,2.02801 0.006,4.05604 0.0196,6.08401 0.16346,25.48324 0.16346,25.48324 -8.11528,35.95993 -6.65018,6.61364 -12.94112,9.45811 -22.375,9.625 -8.28538,-0.14658 -14.12142,-2.12563 -20.27734,-7.67969 -7.3503,-7.68258 -9.17756,-14.90047 -9.1012,-25.24545 -0.005,-1.0119 -0.009,-2.0238 -0.0137,-3.06637 -0.0116,-3.37845 -0.002,-6.7565 0.008,-10.13495 -0.004,-2.42875 -0.01,-4.85749 -0.0163,-7.28623 -0.0117,-5.22946 -0.0145,-10.45884 -0.01,-15.6883 0.007,-7.59785 -0.006,-15.19557 -0.0235,-22.79339 -0.0286,-12.37749 -0.0462,-24.75496 -0.0546,-37.13248 -5.1e-4,-0.74019 -10e-4,-1.48038 -0.002,-2.243 -0.004,-5.24338 -0.007,-10.48676 -0.01,-15.73014 -0.003,-5.9445 -0.007,-11.889 -0.0112,-17.8335 -7.5e-4,-1.09709 -7.5e-4,-1.09709 -0.002,-2.21635 -0.009,-12.30176 -0.0294,-24.60347 -0.058,-36.9052 -0.0155,-6.88345 -0.0248,-13.76681 -0.0186,-20.65028 0.0228,-27.02559 0.0245,-53.81934 -3.05347,-80.69663 C 911.77343,408.89238 911.38261,403.44802 911,398 c -38.17429,1.83947 -38.17429,1.83947 -76,7 -1.06702,0.194 -2.13404,0.38801 -3.2334,0.58789 C 806.02961,410.40069 781.2145,419.16802 757,429 c 0.0704,1.16834 0.1408,2.33668 0.21334,3.54042 0.63717,11.00249 0.9376,21.95734 0.94056,32.97656 0.004,1.67157 0.008,3.34313 0.0133,5.0147 0.0118,4.5326 0.0176,9.06519 0.0222,13.5978 0.006,4.9113 0.0176,9.82258 0.0283,14.73386 0.0244,11.81721 0.0391,23.63441 0.0521,35.45164 0.004,3.34032 0.008,6.68064 0.0118,10.02096 0.0252,20.78015 0.0473,41.5603 0.0597,62.34046 0.003,4.79594 0.006,9.59187 0.009,14.38781 7.4e-4,1.19185 0.001,2.3837 0.002,3.61167 0.0127,19.34962 0.046,38.69916 0.0875,58.04875 0.0424,19.89564 0.066,39.79125 0.0712,59.68694 0.003,11.15937 0.0152,22.31859 0.0481,33.47791 0.028,9.49094 0.039,18.98171 0.0278,28.47268 -0.005,4.83987 -0.002,9.67938 0.0225,14.5192 0.0225,4.43297 0.0217,8.86531 0.003,13.29829 -0.003,1.60055 0.003,3.20114 0.0178,4.80162 0.092,10.45318 -0.96045,18.39336 -7.80587,26.63982 -8.63257,7.33848 -14.76171,8.99498 -26.19922,8.72266 C 716.04402,871.53004 709.73229,867.29451 704,861 c -5.78638,-8.06114 -5.69739,-16.4545 -5.60127,-25.97049 -0.002,-1.61318 -0.007,-3.22636 -0.0129,-4.83953 -0.0111,-4.41285 0.009,-8.82509 0.0339,-13.23786 0.0217,-4.76606 0.0143,-9.53208 0.0105,-14.29817 -0.003,-8.25091 0.013,-16.50161 0.0412,-24.75246 0.0407,-11.92923 0.0537,-23.85836 0.0599,-35.78765 0.0109,-19.35571 0.0442,-38.71131 0.0915,-58.06697 0.0459,-18.79914 0.0813,-37.59826 0.10254,-56.39746 10e-4,-1.15931 0.003,-2.31862 0.004,-3.51307 0.007,-5.8161 0.0129,-11.63219 0.0191,-17.44829 C 698.80019,558.45863 698.88824,510.22932 699,462 c -2.24953,1.60292 -4.49856,3.20653 -6.74731,4.81055 -0.95079,0.67733 -0.95079,0.67733 -1.92079,1.36834 -9.78619,6.98283 -19.08492,14.38514 -27.8944,22.5672 -2.06594,1.91034 -4.16701,3.74721 -6.3125,5.56641 -8.66533,7.53973 -16.74105,15.89786 -24.125,24.6875 -0.80566,0.92813 -1.61133,1.85625 -2.44141,2.8125 -64.26247,74.31113 -87.84653,164.80035 -87.69551,280.50696 0.0181,6.06707 0.0255,12.13414 0.0325,18.20123 0.014,11.99963 0.0355,23.99923 0.0634,35.99884 0.002,0.71281 0.003,1.42562 0.005,2.16003 0.0118,5.06515 0.0239,10.13029 0.036,15.19544 0.0139,5.82454 0.0276,11.64909 0.041,17.47363 0.002,0.71858 0.003,1.43716 0.005,2.17752 0.0274,11.99158 0.0469,23.98316 0.0608,35.97476 0.01,8.01841 0.0296,16.03671 0.0569,24.05509 0.0149,4.97492 0.0215,9.94983 0.0259,14.92477 0.004,2.30221 0.0109,4.60442 0.0219,6.90661 0.0145,3.14599 0.0165,6.29176 0.0156,9.43778 0.007,0.90593 0.0144,1.81185 0.0218,2.74522 -0.0322,9.97626 -1.8983,18.50492 -8.24884,26.42962 -8.56397,7.3126 -17.08708,8.9597 -28.05859,8.2471 -6.86006,-0.8731 -12.66238,-4.99 -16.94145,-10.2471 -3.34243,-4.847 -6.11778,-8.9642 -6.13587,-15.01066 -0.005,-0.81837 -0.01,-1.63674 -0.0147,-2.47991 -3e-4,-0.89889 -5.9e-4,-1.79778 -9e-4,-2.72391 -0.006,-1.4331 -0.006,-1.4331 -0.0128,-2.89514 -0.0131,-3.21672 -0.0191,-6.43341 -0.0251,-9.65015 -0.008,-2.30143 -0.0163,-4.60286 -0.0251,-6.90429 -0.0181,-4.96816 -0.0331,-9.93632 -0.0458,-14.9045 -0.0202,-7.87963 -0.0479,-15.75923 -0.0767,-23.63884 -0.0399,-11.07343 -0.0767,-22.14688 -0.11071,-33.22033 -0.002,-0.70299 -0.004,-1.40598 -0.007,-2.13027 -0.0131,-4.26983 -0.0262,-8.53965 -0.0392,-12.80948 -0.0194,-6.35863 -0.039,-12.71725 -0.0588,-19.07588 -0.002,-0.69485 -0.004,-1.3897 -0.007,-2.10561 -0.0364,-11.66761 -0.0764,-23.33521 -0.11876,-35.00281 -0.0281,-7.80593 -0.0518,-15.61186 -0.0719,-23.41782 -0.0136,-4.90262 -0.031,-9.80522 -0.0493,-14.70783 -0.008,-2.24301 -0.0138,-4.48603 -0.0181,-6.72905 -0.0432,-21.51604 -0.35197,-43.1021 -2.43312,-64.53102 -0.0628,-0.65374 -0.1256,-1.30747 -0.19031,-1.98102 -7.23255,-74.13029 -38.80231,-145.8153 -89.36352,-200.67523 -1.85879,-2.03739 -3.66622,-4.10463 -5.46875,-6.19141 -3.67422,-4.20737 -7.52512,-8.20571 -11.47656,-12.15234 -0.70101,-0.7002 -1.40202,-1.4004 -2.12427,-2.12183 -4.6246,-4.57052 -9.33507,-8.81024 -14.45386,-12.82348 -2.41354,-1.91249 -4.73064,-3.90561 -7.04296,-5.93823 C 335.38999,467.75763 319.32475,457.44815 303,448 c -0.81066,-0.47115 -1.62132,-0.9423 -2.45654,-1.42773 C 241.86833,412.77286 179.93467,401.06676 113,398 c -0.0541,0.81122 -0.10812,1.62245 -0.16382,2.45825 -0.40182,5.84148 -0.83248,11.67458 -1.38285,17.50381 -1.6241,17.28744 -2.66596,34.42784 -2.54049,51.79551 0.007,1.85939 0.0139,3.71878 0.0197,5.57818 0.0125,3.90626 0.0281,7.81249 0.0457,11.71874 0.0363,8.82679 0.0244,17.65353 0.0131,26.48036 -0.004,3.03766 -0.007,6.07532 -0.009,9.11298 -0.017,17.85063 -0.0611,35.70118 -0.11408,53.55174 -0.0349,11.9409 -0.06,23.88169 -0.0635,35.82265 -0.003,10.54475 -0.0235,21.08929 -0.065,31.63396 -0.0212,5.54531 -0.034,11.09032 -0.0256,16.63566 0.008,5.18404 -0.008,10.36748 -0.0419,15.5514 -0.008,1.8938 -0.008,3.78766 0.001,5.68145 0.0546,12.11234 -0.59788,20.88545 -8.9234,30.35031 -7.40888,6.68221 -15.09658,8.02144 -24.86719,7.57031 -7.87402,-0.90305 -14.0638,-4.51659 -19.3012,-10.44122 -6.40351,-8.11402 -6.16334,-17.00949 -6.12751,-26.89447 -0.009,-1.60179 -0.0192,-3.20357 -0.0313,-4.80534 -0.0278,-4.34627 -0.0313,-8.69225 -0.0303,-13.0386 -0.003,-4.61471 -0.0289,-9.2293 -0.052,-13.84394 -0.0406,-8.79876 -0.0609,-17.59745 -0.074,-26.39629 -0.0155,-9.7815 -0.0541,-19.56286 -0.0946,-29.34428 -0.0904,-22.07618 -0.13889,-44.15234 -0.16861,-66.22867 -0.003,-2.39323 -0.007,-4.78646 -0.0103,-7.17969 -0.0141,-9.59484 -0.0233,-19.18968 -0.0282,-28.78453 -9.6e-4,-1.16485 -0.002,-2.32971 -0.003,-3.52986 C 48.90798,408.85548 60.1765,333.85021 95,261 c 0.39623,-0.8416 0.79245,-1.68319 1.20068,-2.55029 18.94515,-40.12525 45.04728,-77.11126 75.17432,-109.60596 2.44738,-2.65132 4.85291,-5.33452 7.25,-8.03125 6.0448,-6.72362 12.36596,-13.0678 19.23437,-18.94922 2.47249,-2.15215 4.86334,-4.38309 7.26563,-6.61328 5.74628,-5.23392 11.78333,-10.00438 17.93994,-14.74316 1.99076,-1.55021 3.9545,-3.13016 5.91553,-4.71778 C 254.16591,75.54982 282.2436,59.55405 311,45 c 0.78246,-0.39929 1.56492,-0.79857 2.37109,-1.20996 C 351.38062,24.53478 393.14985,12.46182 435,5.25 c 1.0634,-0.18659 2.12679,-0.37318 3.22241,-0.56543 22.72424,-3.70072 45.45356,-5.19029 68.45484,-5.07568 z M 326.92139,103.00244 c -1.9973,1.03697 -4.0038,2.05457 -6.01123,3.07178 C 293.38917,120.16425 268.15816,138.59877 245,159 c -0.92813,0.81469 -1.85625,1.62937 -2.8125,2.46875 -6.8307,6.13211 -13.35756,12.56944 -19.875,19.03125 -0.87922,0.86617 -1.75844,1.73234 -2.66431,2.62476 -5.23557,5.22868 -10.05119,10.63992 -14.67627,16.41528 -1.57564,1.96561 -3.17486,3.88676 -4.8352,5.78027 C 192.75718,213.77088 186.32735,222.74445 180,232 c -0.42636,0.61569 -0.85272,1.23138 -1.29199,1.86572 -5.43636,7.85275 -10.45893,15.92111 -15.33301,24.13428 -0.51426,0.8652 -1.02851,1.73041 -1.55835,2.62183 -7.45683,12.69937 -14.04654,25.61529 -19.87915,39.12817 -0.45081,1.03684 -0.45081,1.03684 -0.91074,2.09464 C 132.5873,320.45531 132.5873,320.45531 127,340 c 0.60546,0.0146 1.21091,0.0292 1.83472,0.0442 47.25924,1.24889 91.95831,10.05372 136.59106,25.28393 0.6624,0.2248 1.32481,0.4496 2.00728,0.6812 3.16611,1.09233 6.24266,2.27544 9.29033,3.65767 3.30457,1.58423 3.30457,1.58423 7.27661,1.33301 0.21503,-0.53665 0.43006,-1.07331 0.65161,-1.62622 C 294.94123,343.94267 308.82436,321.5404 327,301 c 0.75926,-0.87012 1.51852,-1.74023 2.30078,-2.63672 C 338.14272,288.35917 347.49036,279.23806 358,271 c 1.45986,-1.18529 1.45986,-1.18529 2.94922,-2.39453 51.15319,-40.75568 118.38753,-56.83806 182.75683,-50.00781 49.54005,6.05114 95.0663,27.60578 131.81739,61.14843 1.67908,1.52813 3.37237,3.04077 5.07812,4.53907 4.44544,3.91494 8.62871,7.85077 12.44141,12.38671 1.52,1.80823 3.06219,3.58029 4.63281,5.34375 16.88108,19.0981 16.88108,19.0981 22.32422,29.98438 -0.899,0.23674 -0.899,0.23674 -1.81616,0.47827 -6.65169,1.85903 -12.60819,4.44222 -18.74634,7.58423 -0.94802,0.47663 -1.89605,0.95326 -2.8728,1.44434 -6.93556,3.53269 -13.66222,7.35337 -20.19776,11.58325 -2.74223,1.71169 -4.79854,2.8877 -8.06616,3.01538 -3.23136,-1.55259 -4.52639,-3.60011 -6.55078,-6.54297 -0.7866,-1.10555 -1.57705,-2.20838 -2.37109,-3.30859 -0.39687,-0.55156 -0.79374,-1.10312 -1.20264,-1.67139 C 655.10228,340.44686 651.5433,336.73129 648,333 c -0.90492,-0.95777 -1.80984,-1.91555 -2.74219,-2.90234 C 610.79038,294.27332 562.98421,275.81952 513.74219,273.87109 482.80633,273.33277 452.09914,281.24141 425,296 c -0.90009,0.48122 -0.90009,0.48122 -1.81836,0.97217 C 398.55145,310.19377 377.6477,328.61909 361,351 c -0.95326,1.2607 -0.95326,1.2607 -1.92578,2.54687 -7.39016,10.12537 -13.19861,21.34608 -18.57422,32.64063 -0.3383,0.70649 -0.6766,1.41297 -1.02515,2.14087 -1.64278,3.52911 -2.93436,6.80564 -3.47485,10.67163 0.93457,0.48727 1.86914,0.97453 2.83203,1.47656 C 350.3513,406.63623 360.659,414.04802 371,422 c 0.65807,0.50354 1.31613,1.00708 1.99414,1.52588 16.50677,12.66527 31.95878,26.36828 46.72855,41.02158 1.42172,1.40981 2.84738,2.81564 4.27316,4.22134 6.10791,6.10027 11.61567,12.49409 17.00415,19.2312 0.93502,1.10163 1.87635,2.19797 2.82813,3.28516 C 472.53528,524.20578 494.98352,563.5342 511,604 c 0.66,0 1.32,0 2,0 0.27038,-0.63164 0.54076,-1.26328 0.81934,-1.91406 C 523.78838,578.88891 534.46684,556.35022 548,535 c 0.41959,-0.66258 0.83918,-1.32516 1.27148,-2.00781 C 559.64244,516.75939 571.22113,501.40493 584,487 c 1.22976,-1.41616 2.45902,-2.83274 3.6875,-4.25 1.10365,-1.25046 2.20782,-2.50045 3.3125,-3.75 0.60199,-0.6948 1.20398,-1.38961 1.82422,-2.10547 5.04267,-5.73423 10.34942,-11.17392 15.73828,-16.58203 0.70818,-0.71277 1.41636,-1.42554 2.146,-2.15991 4.60177,-4.58845 9.3612,-8.92143 14.2915,-13.15259 1.09467,-0.96119 2.18847,-1.92338 3.28125,-2.88672 C 636.86439,434.6146 645.7762,427.69024 655,421 c 0.74991,-0.54914 1.49982,-1.09828 2.27246,-1.66406 C 727.44136,368.13398 810.75307,342.5225 897,340 884.22221,289.31243 854.28433,240.05383 820,201 c -0.63293,-0.73219 -1.26586,-1.46438 -1.91797,-2.21875 C 816.40088,196.84234 814.70249,194.9202 813,193 c -0.63809,-0.7222 -1.27617,-1.44439 -1.93359,-2.18848 -5.62824,-6.31756 -11.46141,-12.38962 -17.44141,-18.37402 -0.86883,-0.88107 -1.73766,-1.76215 -2.63281,-2.66992 -5.26908,-5.28077 -10.71785,-10.14231 -16.54102,-14.80445 -1.95006,-1.56179 -3.85703,-3.14625 -5.73633,-4.79126 C 759.91234,142.49599 750.57364,135.67993 741,129 c -0.64727,-0.45262 -1.29454,-0.90524 -1.96143,-1.37158 C 691.17298,94.37827 635.9495,73.4304 578.6875,63.6875 c -0.71621,-0.12356 -1.43241,-0.24712 -2.17032,-0.37442 -21.13596,-3.60091 -42.28634,-4.72746 -63.70468,-4.68808 -1.12603,5.4e-4 -2.25205,10e-4 -3.4122,0.002 -63.35994,0.13964 -126.16652,15.04673 -182.47891,44.37579 z"
+      fill="currentColor"
+    />
+  </Icon>
+);
+
+export default KibisisIcon;
