+++ src/extension/components/icons/KibisisIcon/index.ts
