+++ src/extension/components/icons/VoiAssetIcon/VoiAssetIcon.tsx
