+++ src/extension/components/icons/VoiAssetIcon/index.ts
