+++ src/extension/components/icons/WarningIcon/WarningIcon.tsx
