+++ src/extension/components/icons/WarningIcon/index.ts
