+++ src/extension/modals/AddAssetsModal/AddAssetsARC0200AssetItem.tsx
@@ -15,7 +15,7 @@ import { IoChevronForward, IoCheckmarkCircleOutline } from 'react-icons/io5';
 // components
 import AssetAvatar from '@extension/components/AssetAvatar';
 import AssetBadge from '@extension/components/AssetBadge';
-import AssetIcon from '@extension/components/AssetIcon';
+import AssetIcon from '@extension/components/icons/AssetIcon';
 
 // constants
 import {
