+++ src/extension/modals/AddAssetsModal/AddAssetsARC0200AssetSummaryModalContent.tsx
@@ -6,7 +6,7 @@ import { useTranslation } from 'react-i18next';
 // components
 import AssetAvatar from '@extension/components/AssetAvatar';
 import AssetBadge from '@extension/components/AssetBadge';
-import AssetIcon from '@extension/components/AssetIcon';
+import AssetIcon from '@extension/components/icons/AssetIcon';
 import CopyIconButton from '@extension/components/CopyIconButton';
 import ModalItem from '@extension/components/ModalItem';
 import ModalTextItem from '@extension/components/ModalTextItem';
