+++ src/extension/modals/AddAssetsModal/AddAssetsStandardAssetItem.tsx
@@ -14,7 +14,7 @@ import { IoCheckmarkCircleOutline, IoChevronForward } from 'react-icons/io5';
 // components
 import AssetAvatar from '@extension/components/AssetAvatar';
 import AssetBadge from '@extension/components/AssetBadge';
-import AssetIcon from '@extension/components/AssetIcon';
+import AssetIcon from '@extension/components/icons/AssetIcon';
 
 // constants
 import {
